Edit done

+++ src/main/java/org/example/kinoxp/repositories/ScreenRepository.java
@@ -3,5 +3,5 @@ package org.example.kinoxp.repositories;
 import org.example.kinoxp.models.Screen;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ScreenRepository extends JpaRepository<Screen, Long> {
+public interface ScreenRepository extends JpaRepository<Screen, Byte> {
 }
